one more

+++ gatsby-node.js
@@ -20,8 +20,10 @@ exports.onPreInit = () => {
 
 exports.onPostBuild = () => {
   fs.renameSync(path.join(__dirname, "public"), path.join(__dirname, "build"));
-  fs.renameSync(
-    path.join(__dirname, "public_dev"),
-    path.join(__dirname, "public")
-  );
+  if (fs.existsSync(path.join(__dirname, "public_dev"))) {
+    fs.renameSync(
+      path.join(__dirname, "public_dev"),
+      path.join(__dirname, "public")
+    );
+  }
 };
